requirements.txt
amqp==5.3.1
anyio==4.12.0
asgiref==3.11.0
asteval==0.9.31
attrs==25.4.0
autobahn==25.12.1
Automat==25.4.16
billiard==4.2.4
cbor2==5.7.1
celery==5.3.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.13.0
cryptography==46.0.3
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory_boy==3.3.3
Faker==38.2.0
gunicorn==21.2.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pymodbus==3.5.0
pyOpenSSL==25.3.0
pyserial==3.5
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
uvicorn==0.27.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.6.0
zope.interface==8.1.1
